Use `ReflectionMethod` without fully qualified namespace in `ActiveQueryTest`.
