+++ tests/Stubs/ActiveQuery/AlternativeActiveQuery.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\ActiveRecord\Tests\Stubs\ActiveQuery;
+
+use Yiisoft\ActiveRecord\ActiveQuery;
+
+final class AlternativeActiveQuery extends ActiveQuery {}
