+++ tests/Stubs/ActiveRecord/CustomerWithDeleteInternalOverride.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\ActiveRecord\Tests\Stubs\ActiveRecord;
+
+final class CustomerWithDeleteInternalOverride extends Customer
+{
+    protected function deleteInternal(): int
+    {
+        return 77;
+    }
+}
